requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
stripe
django-cors-headers
google-auth
python-dotenv
